Mise a jour BDD SQL

+++ app/src/main/java/com/openclassrooms/realestatemanager/dummy/MainActivity.java
@@ -1,9 +1,6 @@
 package com.openclassrooms.realestatemanager.dummy;
 
 import android.Manifest;
-import android.app.AlertDialog;
-import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.location.LocationListener;
@@ -30,17 +27,22 @@ import androidx.recyclerview.widget.RecyclerView;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import com.google.firebase.firestore.FirebaseFirestoreException;
-import com.openclassrooms.realestatemanager.Api.DataBaseSQL;
-import com.openclassrooms.realestatemanager.utils.Adaptateur;
-import com.openclassrooms.realestatemanager.activity.AddInformationActivity;
 import com.openclassrooms.realestatemanager.Api.DI;
+import com.openclassrooms.realestatemanager.Api.DataBaseSQL;
 import com.openclassrooms.realestatemanager.Api.ExtendedServiceEstate;
-import com.openclassrooms.realestatemanager.activity.MapsActivity;
+import com.openclassrooms.realestatemanager.Api.NearbyDao;
 import com.openclassrooms.realestatemanager.R;
+import com.openclassrooms.realestatemanager.activity.AddInformationActivity;
+import com.openclassrooms.realestatemanager.activity.MapsActivity;
 import com.openclassrooms.realestatemanager.activity.SearchActivity;
-import com.openclassrooms.realestatemanager.utils.Utils;
+import com.openclassrooms.realestatemanager.modele.ImagesRealEstate;
+import com.openclassrooms.realestatemanager.modele.NearbyEstate;
 import com.openclassrooms.realestatemanager.modele.RealEstate;
+import com.openclassrooms.realestatemanager.utils.Adaptateur;
+import com.openclassrooms.realestatemanager.utils.Utils;
+import com.openclassrooms.realestatemanager.utils.Utils.CallBackInterfaceForBDD;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -53,6 +55,7 @@ import java.util.List;
  */
 public class MainActivity extends AppCompatActivity {
 
+    int countLoup = 0;
     /**
      * Whether or not the activity is in two-pane mode, i.e. running on a tablet
      * device.
@@ -68,31 +71,12 @@ public class MainActivity extends AppCompatActivity {
     private boolean amIInEuro = true;
     private SwipeRefreshLayout mSwipeRefreshLayout;
     private TextView textNotif;
-    int countLoup=0;
     private DataBaseSQL database;
-
-    private static void buttonInternetInfo(Context context) {
-        AlertDialog alertDialog = buttonInternetInfoDialog(context);
-        alertDialog.show();
-    }
-
-    private static AlertDialog buttonInternetInfoDialog(final Context context) {
-        AlertDialog.Builder builder = new AlertDialog.Builder(context);
-        builder.setMessage(R.string.conexioGPStest).setTitle(R.string.alertinternet).setPositiveButton(R.string.internetActivate, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                Utils.internetOnVerify(context);
-            }
-        });
-        return builder.create();
-    }
 
     @RequiresApi(api = Build.VERSION_CODES.M)
     private void askPermission() {
         if ((MainActivity.this.checkSelfPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED)) {
             ActivityCompat.requestPermissions(MainActivity.this, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, 1);
-        } else {
-            //    placeMeOnMap(googleMap);
         }
     }
 
@@ -113,6 +97,7 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
+
     @RequiresApi(api = Build.VERSION_CODES.M)
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -123,19 +108,59 @@ public class MainActivity extends AppCompatActivity {
         setTitle(R.string.Welcome);
         initiateDataBaseSQL();
         detailsIfTablet();
+        deployementButtonCancel();
+        resultsActivityIfEstateExist();
+        testResultSqlRequet();
+    }
+
+    private void testResultSqlRequet() {
+        List<String> listTest= new ArrayList<>();
+        DataBaseSQL database = DataBaseSQL.getInstance(this);
+        LiveData<List<RealEstate>> datalist = database.estateDao().selectAllEstateSorted(null,null,null,null,
+                null,null,null,null,null,0);
+        datalist.observe(this, new Observer<List<RealEstate>>() {
+            @Override
+            public void onChanged(List<RealEstate> realEstateList) {
+                Toast.makeText(MainActivity.this, "" + realEstateList.size(), Toast.LENGTH_SHORT).show();
+            }
+        });
+    }
+
+
+    @RequiresApi(api = Build.VERSION_CODES.M)
+    private void resultsActivityIfEstateExist() {
+        List<RealEstate> realEstateList = grabEstatFromSearchActivity();
+        if (realEstateList!=null && realEstateList.size() > 0) {
+            Toast.makeText(this, ""+realEstateList.size(), Toast.LENGTH_SHORT).show();
+            adapter = new Adaptateur(Utils.sortedbyPriceDecroissant(realEstateList), mTwoPane, this);
+            recyclerView = findViewById(R.id.RecyclerviewEstate);
+            recyclerView.setHasFixedSize(true);
+            layoutManager = new LinearLayoutManager(MainActivity.this);
+            recyclerView.setLayoutManager(layoutManager);
+            recyclerView.setAdapter(adapter);
+            initiateButtonCAncel().setVisibility(View.VISIBLE);
+
+        } else {
+            NoExistingEstateAction();
+        }
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.M)
+    private void NoExistingEstateAction() {
         saveDataInSQLITE();
-        takeDataInBDDIfInternetIsHere();
+        takeImageInSQLITE();
+        DeploytempHandler();
+//        takeDataInBDDIfInternetIsHere();
         deployementButtonAdd();
         deployementButtonMail();
         onSwipeToRefresh();
         Utils.internetOnVerify(this);
-        DeploytempHandler();
         deployementNotificationMail();
         askPermission();
     }
 
     private void initiateDataBaseSQL() {
-        database=DataBaseSQL.getInstance(this);
+        database = DataBaseSQL.getInstance(this);
     }
 
     private void deployementButtonMail() {
@@ -156,6 +181,15 @@ public class MainActivity extends AppCompatActivity {
         return findViewById(R.id.noInternet);
     }
 
+    private List<RealEstate> grabEstatFromSearchActivity() {
+        List<RealEstate> estates = null;
+        Intent intent = MainActivity.this.getIntent();
+        Bundle extra = intent.getExtras();
+        if (extra != null) {
+            estates = (List<RealEstate>) extra.getSerializable("RealEstate");
+        }
+        return estates;
+    }
 
     private void onSwipeToRefresh() {
         mSwipeRefreshLayout = findViewById(R.id.swipeRefresh);
@@ -170,11 +204,34 @@ public class MainActivity extends AppCompatActivity {
 
     private void takeDataInBDDIfInternetIsHere() {
         if (Utils.internetOnVerify(this)) {
-            Utils.saveDataInBDD(new Utils.CallBackInterfaceForBDD() {
+            Utils.takeDataInBDD(new CallBackInterfaceForBDD() {
                 @Override
-                public void onFinish(List<RealEstate> realEstateList, FirebaseFirestoreException e) {
-                    if (listTemp.size() == 0 || listTempUpdate.size()==0) {
+                public void onFinishEstate(List<RealEstate> realEstateList, FirebaseFirestoreException e) {
+                    if (listTemp.size() == 0 || listTempUpdate.size() == 0 && realEstateList.size() > 0) {
+                        database.nearbyDao().DeleteAllNearby();
+                        database.imageDao().DeleteAllEstate();
                         database.estateDao().DeleteAllEstate();
+
+                        for (int i = 0; i < realEstateList.size(); i++) {
+                            database.estateDao().insertEstate(realEstateList.get(i));
+                        }
+                    } else {
+                        Toast.makeText(MainActivity.this, R.string.actualisation, Toast.LENGTH_SHORT).show();
+                    }
+                }
+
+                @Override
+                public void onFail() {
+                    Toast.makeText(MainActivity.this, R.string.nonew, Toast.LENGTH_SHORT).show();
+                }
+            });
+            Utils.takeDataInBDD(new CallBackInterfaceForBDD() {
+                @Override
+                public void onFinishEstate(List<RealEstate> realEstateList, FirebaseFirestoreException e) {
+                    if (listTemp.size() == 0 || listTempUpdate.size() == 0 && realEstateList.size() > 0) {
+                        database.imageDao().DeleteAllEstate();
+                        database.estateDao().DeleteAllEstate();
+
                         for (int i = 0; i < realEstateList.size(); i++) {
                             database.estateDao().insertEstate(realEstateList.get(i));
                         }
@@ -214,9 +271,41 @@ public class MainActivity extends AppCompatActivity {
         });
     }
 
+    private void takeImageInSQLITE() {
+        DataBaseSQL database = DataBaseSQL.getInstance(this);
+        LiveData<List<ImagesRealEstate>> datalist = database.imageDao().selectAllImage();
+        LiveData<List<NearbyEstate>> datalist3 = database.nearbyDao().selectAllImage();
+        LiveData<List<ImagesRealEstate>> datalist2 = database.imageDao().selectAllImageDeuxFois(3);
+        datalist.observe(this, new Observer<List<ImagesRealEstate>>() {
+            @Override
+            public void onChanged(List<ImagesRealEstate> imagesRealEstates) {
+                if (imagesRealEstates.size() > 0) {
+//                    Toast.makeText(MainActivity.this, ""+imagesRealEstates.size(), Toast.LENGTH_SHORT).show();
+                }
+            }
+        });
+        datalist2.observe(this, new Observer<List<ImagesRealEstate>>() {
+            @Override
+            public void onChanged(List<ImagesRealEstate> imagesRealEstates) {
+                if (imagesRealEstates.size() > 0) {
+                }
+            }
+        });
+        datalist3.observe(this, new Observer<List<NearbyEstate>>() {
+            @Override
+            public void onChanged(List<NearbyEstate> nearbyEstates) {
+                if (nearbyEstates.size() > 0) {
+//                    Toast.makeText(MainActivity.this, ""+nearbyEstates.size(), Toast.LENGTH_SHORT).show();
+                }
+            }
+        });
+    }
+
+
+
     private void searchForNotif(List<RealEstate> realEstates) {
-        countLoup=countLoup+1;
-        if (countLoup==1) {
+        countLoup = countLoup + 1;
+        if (countLoup == 1) {
             listTemp.clear();
             for (int i = 0; i < realEstates.size(); i++) {
                 if (realEstates.get(i).getTempInsert().equals("true")) {
@@ -242,7 +331,9 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private void majNotif() {
-        textNotif.setText(String.valueOf(listTemp.size()+listTempUpdate.size()));
+        if (textNotif!=null) {
+            textNotif.setText(String.valueOf(listTemp.size() + listTempUpdate.size()));
+        }
     }
 
     private void DeploytempHandler() {
@@ -256,6 +347,7 @@ public class MainActivity extends AppCompatActivity {
         } else {
             if (Utils.internetOnVerify(this)) {
                 if (listTemp.size() > 0) {
+
                     for (int i = 0; i < listTemp.size(); i++) {
                         Utils.sendItToMyBDDatRealEstate(listTemp.get(i));
                         try {
@@ -272,6 +364,9 @@ public class MainActivity extends AppCompatActivity {
                         database.estateDao().upDateEstate(listTemp.get(i));
                     }
                     Toast.makeText(this, R.string.sent, Toast.LENGTH_SHORT).show();
+                    if (listTempUpdate.size() > 0) {
+                        sendTempUpdateFileToFireB();
+                    }
                     listTemp.clear();
                     majNotif();
                 }
@@ -288,17 +383,7 @@ public class MainActivity extends AppCompatActivity {
             if (Utils.internetOnVerify(this)) {
                 if (listTempUpdate.size() > 0) {
                     for (int i = 0; i < listTempUpdate.size(); i++) {
-                        Utils.upDateMyBDDPlease(listTempUpdate.get(i),listTempUpdate.get(i));
-//                        try {
-//                            Utils.uploadImage(listTempUpdate.get(i), this, new Utils.CallBackImage() {
-//                                @Override
-//                                public void onFinish(List<String> s) {
-//
-//                                }
-//                            });
-//                        } catch (Exception e) {
-//                            e.printStackTrace();
-//                        }
+                        Utils.upDateMyBDDPlease(listTempUpdate.get(i), listTempUpdate.get(i));
                         listTempUpdate.get(i).setTempInsert("False");
                         database.estateDao().upDateEstate(listTempUpdate.get(i));
                     }
@@ -316,6 +401,26 @@ public class MainActivity extends AppCompatActivity {
         if (findViewById(R.id.item_detail_container) != null) {
             mTwoPane = true;
         }
+    }
+    private void deployementButtonCancel() {
+        final ImageButton cancelButton = initiateButtonCAncel();
+        activateButtonCancel(cancelButton);
+    }
+
+    private void activateButtonCancel(final ImageButton cancelButton) {
+        cancelButton.setOnClickListener(new View.OnClickListener() {
+            @RequiresApi(api = Build.VERSION_CODES.M)
+            @Override
+            public void onClick(View view) {
+                NoExistingEstateAction();
+                cancelButton.setVisibility(View.GONE);
+            }
+        });
+    }
+
+    private ImageButton initiateButtonCAncel() {
+        return findViewById(R.id.buttoncancelactivitymain);
+
     }
 
     private void deployementButtonAdd() {
@@ -384,7 +489,7 @@ public class MainActivity extends AppCompatActivity {
                     Intent intent = new Intent(MainActivity.this, MapsActivity.class);
                     startActivity(intent);
                 } else {
-                    Toast.makeText(MainActivity.this, getString(R.string.providerinternet) , Toast.LENGTH_SHORT).show();
+                    Toast.makeText(MainActivity.this, getString(R.string.providerinternet), Toast.LENGTH_SHORT).show();
                 }
             }
         });
@@ -418,7 +523,7 @@ public class MainActivity extends AppCompatActivity {
     private void convertDollarList() {
         for (int i = 0; i < listRealEstate.size(); i++) {
             int converted = Utils.convertDollarToEuro(Integer.valueOf(listRealEstate.get(i).getPrix()));
-            listRealEstate.get(i).setPrix(String.valueOf((converted)));
+            listRealEstate.get(i).setPrix(converted);
         }
         deployRecyclerView();
     }
@@ -426,7 +531,7 @@ public class MainActivity extends AppCompatActivity {
     private void convertEuroList() {
         for (int i = 0; i < listRealEstate.size(); i++) {
             int converted = Utils.convertEuroToDollar(Integer.valueOf(listRealEstate.get(i).getPrix()));
-            listRealEstate.get(i).setPrix(String.valueOf((converted)));
+            listRealEstate.get(i).setPrix(converted);
         }
         deployRecyclerView();
     }
@@ -434,8 +539,13 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onResume() {
         super.onResume();
+        //no();
+    }
+
+    private void no() {
         takeDataInBDDIfInternetIsHere();
         saveDataInSQLITE();
+        takeImageInSQLITE();
         majNotif();
     }
 }
